api/requirements.txt
selenium
python-dotenv
mysql-connector-python
fastapi
uvicorn
pydantic 